Update README.md
We started by creating a new React project using Vite, a modern and fast build tool. During the setup, we selected React with JavaScript as the framework. After setting up the project structure, we installed the necessary packages and dependencies to run the app smoothly.

Once the development environment was ready, we installed additional libraries like React Router for navigation and React Icons for using icons easily. With everything in place, we began building a fully responsive portfolio from scratch using clean CSS